requirements.txt
# Core LLM & LangChain
langchain>=0.3.7
langchain-community>=0.3.7
langchain-core>=0.3.19
langchain-groq>=0.2.1
langchain-huggingface>=0.1.0
langchain-ollama>=0.2.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.66
langgraph-checkpoint-sqlite==3.0.3
aiosqlite>=0.20.0
sqlite-vec>=0.1.3

# Vector Store & Embeddings
# FIX: added langchain-chroma (replaces deprecated langchain_community.vectorstores.Chroma)
langchain-chroma>=0.1.4
chromadb>=1.5.6
sentence-transformers>=3.3.1
rank-bm25>=0.2.2

# Document Processing
pypdf>=5.1.0
python-docx>=1.1.2
# FIX: added unstructured for UnstructuredWordDocumentLoader (DOCX ingestion)
unstructured>=0.16.0
unstructured[docx]>=0.16.0

# API & Web
fastapi>=0.115.5
uvicorn[standard]>=0.32.1
python-multipart>=0.0.12
httpx>=0.27.2

# Search
tavily-python>=0.5.0

# Frontend
streamlit>=1.40.2
gradio>=5.6.0

# Utilities
pydantic>=2.10.1
pydantic-settings>=2.6.1
python-dotenv>=1.0.1
numpy>=2.0.0
pandas>=2.2.3

# Reranking
flashrank>=0.2.9

# Ollama (optional local LLM)
ollama>=0.4.3
